+++ molo/core/models.py
@@ -445,6 +445,7 @@ class Tag(TranslatablePageMixin, Page):
 
     feature_in_homepage = models.BooleanField(default=False)
 
+
 Tag.promote_panels = [
     FieldPanel('feature_in_homepage'),
     MultiFieldPanel(
@@ -504,12 +505,12 @@ class Main(CommentedPageMixin, Page):
     subpage_types = []
 
     def bannerpages(self):
-        return BannerPage.objects.child_of(self).filter(
+        index_page = BannerIndexPage.objects.child_of(self).live().first()
+        return BannerPage.objects.child_of(index_page).filter(
             languages__language__is_main_language=True).specific()
 
     def sections(self):
-        index_page = SectionIndexPage.objects.child_of(
-            self).live().first()
+        index_page = SectionIndexPage.objects.child_of(self).live().first()
         return SectionPage.objects.child_of(index_page).filter(
             languages__language__is_main_language=True).specific()
 
@@ -677,9 +678,7 @@ class SectionIndexPage(CommentedPageMixin, Page, PreventDeleteMixin):
     commenting_close_time = models.DateTimeField(null=True, blank=True)
 
     def celery_copy(self, *args, **kwargs):
-        site = kwargs['to'].get_site()
-        main = site.root_page
-        SectionIndexPage.objects.child_of(main).delete()
+        SectionIndexPage.objects.child_of(kwargs['to']).delete()
         return super(SectionIndexPage, self).copy(*args, **kwargs)
 
     def copy(self, *args, **kwargs):
@@ -885,7 +884,8 @@ SectionPage.content_panels = [
             FieldPanel('commenting_open_time'),
             FieldPanel('commenting_close_time'),
         ],
-        heading="Commenting Settings", )
+        heading="Commenting Settings", ),
+    InlinePanel('section_tags', label="Tags for Navigation"),
 ]
 
 SectionPage.settings_panels = [
@@ -1059,6 +1059,14 @@ class ArticlePage(CommentedPageMixin, TranslatablePageMixin, Page):
     def get_absolute_url(self):  # pragma: no cover
         return self.url
 
+    def get_effective_image(self):
+        if self.image:
+            return self.image
+        page = self.get_main_language_page()
+        if page.specific.image:
+            return page.specific.get_effective_image()
+        return ''
+
     def get_parent_section(self):
         return SectionPage.objects.all().ancestor_of(self).last()
 
@@ -1158,6 +1166,19 @@ def demote_featured_articles(sender, instance, **kwargs):
         instance.featured_in_section_start_date is None and \
             instance.featured_in_section is True:
         instance.featured_in_section = False
+
+
+class SectionPageTags(Orderable):
+    page = ParentalKey(SectionPage, related_name='section_tags')
+    tag = models.ForeignKey(
+        'wagtailcore.Page',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+        help_text=_('Tags for tag navigation')
+    )
+    panels = [PageChooserPanel('tag', 'core.Tag')]
 
 
 class ArticlePageTags(Orderable):