+++ molo/core/templatetags/core_tags.py
@@ -5,8 +5,9 @@ from django.utils.safestring import mark_safe
 from django.db.models import Case, When
 from markdown import markdown
 
-from molo.core.models import (Page, ArticlePage, SectionPage,
-                              SiteSettings, Languages, Tag, ArticlePageTags)
+from molo.core.models import (
+    Page, ArticlePage, SectionPage, SiteSettings, Languages, Tag,
+    ArticlePageTags, SectionIndexPage)
 
 register = template.Library()
 
@@ -74,6 +75,14 @@ def get_translation(context, page):
         return page.get_translation_for(locale_code, context['request'].site)
     else:
         return page
+
+
+@register.assignment_tag(takes_context=True)
+def get_parent(context, page):
+    parent = page.get_parent()
+    if not parent.specific_class == SectionIndexPage:
+        return get_translation(context, parent.specific)
+    return None
 
 
 @register.inclusion_tag(
@@ -147,7 +156,7 @@ def topic_of_the_day(context):
 
 
 @register.inclusion_tag('core/tags/bannerpages.html', takes_context=True)
-def bannerpages(context):
+def bannerpages(context, position=None):
     request = context['request']
     locale = context.get('locale_code')
 
@@ -156,6 +165,18 @@ def bannerpages(context):
     else:
         pages = []
 
+    if position >= 0:
+        banners = get_pages(context, pages, locale)
+        if position > (len(banners) - 1):
+            return None
+        banners = get_pages(context, pages, locale)
+        if banners and len(banners) > position:
+            return {
+                'bannerpages': [banners[position]],
+                'request': context['request'],
+                'locale_code': locale,
+            }
+        return None
     return {
         'bannerpages': get_pages(context, pages, locale),
         'request': context['request'],
@@ -298,18 +319,62 @@ def load_child_articles_for_section(context, section, count=5):
     return articles
 
 
+def get_articles_for_tags_with_translations(
+        request, tag, exclude_list, locale, context, exclude_pks):
+
+    pks = [article_tag.page.pk for article_tag in
+           ArticlePageTags.objects.filter(tag=tag)]
+    return get_pages(
+        context, ArticlePage.objects.descendant_of(
+            request.site.root_page).filter(pk__in=pks).exclude(
+                pk__in=exclude_pks), locale)
+
+
+@register.assignment_tag(takes_context=True)
+def get_tags_for_section(context, section, tag_count=2, tag_article_count=4):
+    request = context['request']
+    locale = context.get('locale_code')
+    # Featured Tags
+    exclude_pks = []
+    tags_list = []
+    main_language_page = section.get_main_language_page()
+    child_articles = ArticlePage.objects.descendant_of(
+        main_language_page).filter(languages__language__is_main_language=True)
+    for article in child_articles:
+        exclude_pks.append(article.pk)
+    related_articles = ArticlePage.objects.filter(
+        related_sections__section__slug=main_language_page.slug)
+    for article in related_articles:
+        exclude_pks.append(article.pk)
+
+    tags = [
+        section_tag.tag.pk for section_tag in
+        section.get_main_language_page().specific.section_tags.all()
+        if section_tag.tag]
+    if tags and request.site:
+        qs = Tag.objects.descendant_of(
+            request.site.root_page).live().filter(pk__in=tags)
+        for tag in qs:
+            tag_articles = get_articles_for_tags_with_translations(
+                request, tag, exclude_pks, locale, context,
+                exclude_pks)
+            if len(tag_articles) > tag_article_count:
+                tag_articles = tag_articles[:tag_article_count]
+            exclude_pks += [p.pk for p in tag_articles]
+            tags = get_pages(context, qs.filter(pk=tag.pk), locale)
+            if tags:
+                tags_list.append((
+                    tags[0],
+                    tag_articles))
+    else:
+        return []
+
+    return tags_list
+
+
 @register.simple_tag(takes_context=True)
 def get_tag_articles(
-        context, latest_articles_count=12, section_count=1, tag_count=4,
-        sec_articles_count=4):
-    def get_positional_tag_articles(request, tag, exclude_list):
-
-        pks = [article_tag.page.pk for article_tag in
-               ArticlePageTags.objects.filter(tag=tag)]
-        return get_pages(
-            context, ArticlePage.objects.descendant_of(
-                request.site.root_page).filter(pk__in=pks).exclude(
-                    pk__in=exclude_pks), locale)
+        context, section_count=1, tag_count=4, sec_articles_count=4):
 
     request = context['request']
     locale = context.get('locale_code')
@@ -319,16 +384,7 @@ def get_tag_articles(
     tags_list = []
     sections_list = []
 
-    # Latest Articles
-    latest_articles = request.site.root_page.specific.latest_articles(
-    )[:latest_articles_count]
-    exclude_pks += [p.pk for p in latest_articles]
-    data.update({
-        'latest_articles': get_pages(
-            context, ArticlePage.objects.filter(
-                pk__in=[p.pk for p in latest_articles]), locale)})
-
-    # Featured Section
+    # Featured Section/s
     sections = request.site.root_page.specific.sections()
     for section in sections[:section_count]:
         sec_articles = ArticlePage.objects.descendant_of(section).filter(
@@ -337,20 +393,51 @@ def get_tag_articles(
                 '-featured_in_homepage_start_date').exclude(
                 pk__in=exclude_pks)
         exclude_pks += [p.pk for p in sec_articles[:sec_articles_count]]
-        sections_list.append((
-            section,
-            get_pages(context, sec_articles, locale)[:sec_articles_count]))
+        section = SectionPage.objects.filter(pk=section.pk)
+        section_for_locale = get_pages(context, section, locale)
+        if section_for_locale:
+            section = section_for_locale[0]
+        sec_articles_for_locale = get_pages(context, sec_articles, locale)
+        if sec_articles_for_locale and len(
+                sec_articles_for_locale) > sec_articles_count:
+            sec_articles_for_locale = sec_articles_for_locale[
+                :sec_articles_count]
+        if section_for_locale:
+            sections_list.append((
+                section,
+                sec_articles_for_locale))
     data.update({'sections': sections_list})
 
-    # Featured Tags
-    for tag in Tag.objects.descendant_of(request.site.root_page).filter(
-            feature_in_homepage=True).live():
-        tag_articles = get_positional_tag_articles(
-            request, tag, exclude_pks)[:tag_count]
+    # Featured Tag/s
+    tag_qs = Tag.objects.descendant_of(request.site.root_page).filter(
+        feature_in_homepage=True).live()
+    for tag in tag_qs:
+        tag_articles = get_articles_for_tags_with_translations(
+            request, tag, exclude_pks, locale,
+            context, exclude_pks)
+        if tag_articles and len(tag_articles) > tag_count:
+            tag_articles = tag_articles[:tag_count]
         exclude_pks += [p.pk for p in tag_articles]
-        tags_list.append((tag, tag_articles))
-
+        tag_for_locale = get_pages(context, tag_qs.filter(pk=tag.pk), locale)
+        if tag_for_locale:
+            tags_list.append((
+                tag_for_locale[0],
+                tag_articles))
+        else:
+            tags_list.append((
+                tag,
+                tag_articles))
     data.update({'tags_list': tags_list})
+
+    # Latest Articles
+    data.update({
+        'latest_articles': get_pages(
+            context, ArticlePage.objects.descendant_of(
+                request.site.root_page).filter(
+                languages__language__is_main_language=True).exact_type(
+                    ArticlePage).exclude(pk__in=exclude_pks).order_by(
+                        '-featured_in_latest'), locale)})
+
     return data