+++ molo/core/views.py
@@ -22,11 +22,13 @@ from wagtail.wagtailsearch.models import Query
 from molo.core.utils import generate_slug, get_locale_code, update_media_file
 from molo.core.models import (
     PageTranslation, ArticlePage, Languages, SiteSettings, Tag,
-    ArticlePageTags)
+    ArticlePageTags, SectionPage)
 from molo.core.templatetags.core_tags import get_pages
 from molo.core.known_plugins import known_plugins
 from molo.core.forms import MediaForm
 from django.views.generic import ListView
+
+from el_pagination.decorators import page_template
 
 
 def csrf_failure(request, reason=""):
@@ -44,7 +46,7 @@ def search(request, results_per_page=10):
 
         results = ArticlePage.objects.descendant_of(main).filter(
             languages__language__locale=locale
-        ).values_list('pk', flat=True)
+        ).exact_type(ArticlePage).values_list('pk', flat=True)
 
         # Elasticsearch backend doesn't support filtering
         # on related fields, at the moment.
@@ -187,18 +189,26 @@ class TagsListView(ListView):
         locale = self.request.LANGUAGE_CODE
 
         if site_settings.enable_tag_navigation:
-            tag = Tag.objects.get(slug=tag)
+            tag = Tag.objects.filter(slug=tag).descendant_of(main).first()
             articles = []
             for article_tag in ArticlePageTags.objects.filter(
                     tag=tag.get_main_language_page()).all():
                 articles.append(article_tag.page.pk)
             articles = ArticlePage.objects.filter(
-                pk__in=articles).order_by(
+                pk__in=articles).descendant_of(main).order_by(
                     'latest_revision_created_at')
-            return get_pages(context, articles, locale)[:count]
+            # count = articles.count() if articles.count() < count else count
+            return get_pages(context, articles[:count], locale)
         return ArticlePage.objects.descendant_of(main).filter(
             tags__name__in=[tag]).order_by(
                 'latest_revision_created_at')
+
+    def get_context_data(self, *args, **kwargs):
+        context = super(TagsListView, self).get_context_data(*args, **kwargs)
+        tag = self.kwargs['tag_name']
+        context.update({'tag': Tag.objects.filter(
+            slug=tag).descendant_of(self.request.site.root_page).first()})
+        return context
 
 
 @user_passes_test(lambda u: u.is_superuser)
@@ -250,3 +260,36 @@ def download_file(request):
                           'django_admin/transfer_media_message.html',
                           {'error_message':
                            'media file does not exist'})
+
+
+@page_template(
+    'patterns/basics/article-teasers/latest-promoted_variations/'
+    'latest-articles_for-paging.html')
+def home_index(
+        request,
+        extra_context=None,
+        template=(
+            'patterns/components/article-teasers/latest-promoted_variations/'
+            'article_for_paging.html')):
+    return render(request, template, {})
+
+
+@page_template(
+    'patterns/basics/sections/sectionpage-article-list-'
+    'standard_for-paging.html')
+def section_index(
+        request,
+        extra_context=None,
+        template=(
+            'patterns/basics/sections/sectionpage-article-list-'
+            'standard_for-paging.html')):
+    section = SectionPage.objects.get(pk=request.GET.get('section'))
+    return render(request, template, {'section': section})
+
+
+@page_template(
+    'patterns/basics/article-teasers/latest-promoted_variations/late'
+    'st-articles_for-feature.html')
+def home_more(
+        request, template='core/main-feature-more.html', extra_context=None):
+    return render(request, template, {})