+++ molo/core/wagtail_hooks.py
@@ -1,6 +1,7 @@
 from django.conf.urls import url
 
-from molo.core.models import LanguageRelation, PageTranslation, Languages
+from molo.core.models import LanguageRelation, PageTranslation, Languages, \
+    ArticlePage, Tag, ArticlePageTags
 
 from django.core import urlresolvers
 from django.utils.translation import ugettext_lazy as _
@@ -46,7 +47,24 @@ def show_main_language_only(parent_page, pages, request):
 
 
 @hooks.register('after_copy_page')
+def add_new_tag_article_relations(request, page, new_page):
+    if new_page.depth <= 2:
+        for article in ArticlePage.objects.descendant_of(new_page):
+            tag_relations = ArticlePageTags.objects.filter(page=article)
+            for relation in tag_relations:
+                new_tag = Tag.objects.descendant_of(
+                    new_page).filter(slug=relation.tag.slug).first()
+                relation.tag = new_tag
+                relation.save()
+
+
+@hooks.register('after_copy_page')
 def copy_translation_pages(request, page, new_page):
+
+    # Only copy translations for TranslatablePageMixin
+    if not hasattr(page.specific, 'copy_language'):
+        return 'Not translatable page'
+
     current_site = page.get_site()
     destination_site = new_page.get_site()
     if current_site is not destination_site and (page.depth > 2):